Updated javascrip[ts
removed ajax references and centralized apiCall.

Decided to just go with react to host the front end. If this thing scales as large as i think it will, it'll require  that it's absteract enough to handle what ever we can throw at it. Switched temp code over to bootsrtapped version. (not the blue looking one...it was really...blue

removed some extra functionality. Large sections will be shifted and moved around. popped most of it into modals used in react

tried to get rid of repetitive code. Still has a few kinks, but we're right around the corner.